Remove continuation_indent_size from .editorconfig
The continuation_indent_size was specified here to ensure that
continuations in conditional clauses did not align with the body of
those code blocks, however this also results in continuations in the
code block body being indented by 8 spaces, which we do not want.

Going forward, 8-space continuations in conditional clauses will be
enforced during Pull Requests, but in general, continuations should be 4
spaces by default.